refactor: MVC 구조에 맞게 리팩터링
